Ticket [https://trac.osgeo.org/fdo/ticket/475]: Build: Fix Warnings and Compile Errors
git-svn-id: http://svn.osgeo.org/fdo/trunk@4480 d98eaee6-4212-0410-900c-dcc49bf69b99

+++ Providers/SHP/Src/UnitTest/InsertTests2.cpp
@@ -46,7 +46,7 @@ void InsertTests2::CreateConnection( Context& context, FdoBoolean recreateDb )
     if (FdoCommonFile::FileExists (dirName)) {
         std::vector<std::wstring> files;
         FdoCommonFile::GetAllFiles (dirName, files );
-        for ( int i = 0; i < files.size(); i++ ) 
+        for ( size_t i = 0; i < files.size(); i++ ) 
             FdoCommonFile::Delete( 
                 dirName + 
 #ifdef _WIN32
